--- nx2/public/plugins/quick-edit/quick-edit.css
@@ -1,310 +1,0 @@
-.remote-cursor-indicator::before {
-  content: attr(data-cursor-remote);
-  position: absolute;
-  background: var(--cursor-remote-color, red);
-  color: white;
-  padding: 2px 6px;
-  border-radius: 3px;
-  font-size: 12px;
-  font-weight: bold;
-  transform: translateY(-100%);
-  margin-top: -4px;
-  white-space: nowrap;
-  z-index: 1000;
-}
-
-.remote-cursor-indicator {
-  position: relative;
-  outline: 2px solid;
-}
-
-/* Hover outline for instrumented editables and blocks (data-prose-index, data-block-index) */
-[data-prose-index]:hover,
-[data-block-index]:hover {
-  outline: 1px solid var(--color-blue-600, blue);
-}
-
-[data-prose-index]:has(.ProseMirror-focused) {
-  outline: 2px solid var(--color-blue-600, blue);
-
-  .ProseMirror-focused {
-    outline: none;
-  }
-}
-
-/* Image drop zone styles */
-picture.image-drop-target {
-  position: relative;
-  display: block;
-}
-
-picture.image-drop-target::before {
-  content: "";
-  position: absolute;
-  inset: 0;
-  border: 3px dashed var(--color-blue-600, #3b82f6);
-  border-radius: 4px;
-  pointer-events: none;
-  z-index: 10;
-}
-
-picture.image-drop-target::after {
-  content: "Drop to replace";
-  position: absolute;
-  inset: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: rgb(59 130 246 / 40%);
-  color: white;
-  font-weight: 600;
-  font-size: 14px;
-  text-shadow: 0 1px 2px rgb(0 0 0 / 30%);
-  pointer-events: none;
-  z-index: 11;
-  border-radius: 4px;
-}
-
-main picture img {
-  cursor: grab;
-}
-
-picture.image-uploading img {
-  opacity: 0.5;
-  filter: grayscale(0.3);
-}
-
-main picture img:hover {
-  outline: 1px solid var(--color-blue-600, blue);
-}
-
-/* Loading state during upload */
-picture.image-uploading {
-  position: relative;
-}
-
-picture.image-uploading::after {
-  content: "";
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  width: 32px;
-  height: 32px;
-  margin: -16px 0 0 -16px;
-  border: 3px solid rgb(59 130 246 / 30%);
-  border-top-color: var(--color-blue-600, #3b82f6);
-  border-radius: 50%;
-  animation: spin 0.8s linear infinite;
-  z-index: 10;
-}
-
-@keyframes spin {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.quick-edit-buttons {
-  position: fixed;
-  bottom: 20px;
-  right: 20px;
-  z-index: 9999;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.quick-edit-close {
-  width: 32px;
-  height: 32px;
-  padding: 0;
-  background: var(--color-blue-500, rgb(19 115 230));
-  border: none;
-  border-radius: 50%;
-  cursor: pointer;
-  box-shadow: 0 2px 8px rgb(0 0 0 / 20%);
-  transition: background 0.2s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.quick-edit-close:hover {
-  background: var(--color-blue-600, #3b82f6);
-}
-
-.quick-edit-close .icon-close {
-  display: block;
-  width: 16px;
-  height: 16px;
-  background-color: white;
-  mask: url("/nx/img/icons/S2IconClose20N-icon.svg") no-repeat center / contain;
-  transition: transform 0.3s ease;
-}
-
-.quick-edit-close.toggled .icon-close {
-  transform: rotate(45deg);
-}
-
-.quick-edit-exit,
-.quick-edit-preview {
-  height: 32px;
-  padding: 0 16px;
-  background: white;
-  border: 1px solid var(--color-blue-500, rgb(19 115 230));
-  color: var(--color-blue-500, rgb(19 115 230));
-  border-radius: 16px;
-  cursor: pointer;
-  box-shadow: 0 2px 8px rgb(0 0 0 / 20%);
-  transition: all 0.2s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 14px;
-  font-weight: 500;
-  white-space: nowrap;
-}
-
-.quick-edit-preview {
-  background: var(--color-blue-500, rgb(19 115 230));
-  color: white;
-}
-
-.quick-edit-exit:hover {
-  background: var(--color-blue-50, #eff6ff);
-  border-color: var(--color-blue-600, #3b82f6);
-  color: var(--color-blue-600, #3b82f6);
-}
-
-.quick-edit-preview:hover {
-  background: var(--color-blue-600, #3b82f6);
-}
-
-/* Floating Toolbar Styles */
-.prosemirror-floating-toolbar {
-  color-scheme: light;
-  font-size: 13px;
-  position: fixed;
-  background: white;
-  border: 1px solid #ccc;
-  border-radius: 8px;
-  padding: 8px;
-  display: none;
-  z-index: 1000;
-  box-shadow: 0 2px 8px rgb(0 0 0 / 15%);
-  color: black;
-}
-
-.prosemirror-floating-toolbar .toolbar-btn {
-  padding: 6px 12px;
-  margin: 0 2px;
-  border: 1px solid #ccc;
-  background: white;
-  cursor: pointer;
-  border-radius: 6px;
-  transition: background 0.2s ease;
-}
-
-.ProseMirror-menuitem {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 2px;
-  border: 1px solid #ccc;
-  background: white;
-  cursor: pointer;
-  border-radius: 6px;
-  transition: background 0.2s ease;
-}
-
-.prosemirror-floating-toolbar .toolbar-btn:hover,
-.ProseMirror-menuitem:hover {
-  background: #f0f0f0;
-}
-
-.prosemirror-floating-toolbar .toolbar-btn.active {
-  background: #e0e0e0;
-  border-color: #999;
-}
-
-.prosemirror-floating-toolbar .toolbar-btn-bold {
-  font-weight: bold;
-}
-
-.prosemirror-floating-toolbar .toolbar-btn-italic {
-  font-style: italic;
-}
-
-.prosemirror-floating-toolbar .toolbar-btn-underline {
-  text-decoration: underline;
-}
-
-.ProseMirror-menuitem > div {
-  padding: 6px 12px;
-}
-
-.ProseMirror-menuitem:has(.ProseMirror-menu-disabled) {
-  display: none;
-}
-
-.da-palettes:has(da-palette) {
-  color-scheme: light;
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgb(0 0 0 / 25%);
-  z-index: 1000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-da-palette {
-  background: white;
-  padding: 30px;
-  border-radius: 8px;
-  box-shadow: 0 0 10px 0 rgb(0 0 0 / 10%);
-}
-
-da-palette .da-palette-field {
-  color: black;
-}
-
-/* Anchor for absolutely positioned quick-edit UI (document space, scrolls with content) */
-html {
-  position: relative;
-}
-
-/* Add to context overlay (instrumented elements hover) */
-.quick-edit-add-to-context-overlay {
-  position: absolute;
-  z-index: 10000;
-  pointer-events: auto;
-}
-
-.quick-edit-add-to-context-btn {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 6px;
-  background: var(--color-blue-600, #2563eb);
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  box-shadow: 0 1px 3px rgb(0 0 0 / 20%);
-}
-
-.quick-edit-add-to-context-btn-icon {
-  display: block;
-  width: 20px;
-  height: 20px;
-  
-  /* SVG uses dark fills; invert for contrast on blue button */
-  filter: brightness(0) invert(1);
-}
-
-.quick-edit-add-to-context-btn:hover {
-  background: var(--color-blue-700, #1d4ed8);
-}
